pyproject.toml
[tool.poetry]
authors = ["thirdweb <sdk@thirdweb.com>"]
description = ""
name = "thirdweb-sdk"
packages = [
  {include = "thirdweb"},
]
readme = "README.md"
version = "2.0.0"

[tool.poetry.dependencies]
dacite = "^1.6.0"
mypy-extensions = "^0.4.3"
python = ">3.7.1,<=3.10.4"
thirdweb-contract-wrappers = "^2.0.4"
web3 = "5.27.0"
thirdweb-eth-account = "^0.6.6"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
mypy = "^0.941"
pylint = "^2.12.2"
types-requests = "^2.27.14"
Sphinx = "^4.5.0"
sphinx-markdown-builder = "^0.5.5"
sphinx-autodoc-typehints = "^1.17.0"
mkdocs = "^1.3.0"
mkdocs-autorefs = "^0.4.1"
mkdocs-material = "^8.2.9"
mkdocs-material-extensions = "^1.0.3"
mkdocstrings = "^0.18.1"
pymdown-extensions = "^9.3"
eth-brownie = "^1.18.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
